Add extraction of categories from images
Starting to develop the code that will allow us to extract the
different categories/labels from each of the images.

This is currently only being tested on two images for speed in
the event of errors. Further dicussion is needed around how
images will be stored after extraction.

Drivable Area and Lane are being excluded from this extraction as
they are not required for this CA.

.gitignore updated to prevent tracking of ./idea.