Add function to generalise image extraction
This function was created to promote DRY programming as it
generalises the image extraction process for training, validation
and test.

The images extracted via this function will store images in a
master directory, training, validation or test and then sub
directories broken down by class (plane, person, car).

Note this function can throw an OSError as we do not want to
overwrite the contents of the users directory if a directory
using the same name already exists.