Refactor image extraction feature
The original code written to extract the class images was tested
on a small sample of images from the Berkley dataset. When I
attempted to run the original code against all 100k images an
issue occurred. The issue appeared to be a rounding issue as the
coordinates given in the the label files are floating point
numbers this lead to problems such as x1=5.6 becoming 6 and
x2=5.8 becoming 6. When the save function was called an exception
was thrown stating (in simple terms) that there was no image to
save as the distance between x1 and x2 was 0.

This lead to the import of PIL Image as I originally hoped this
issue was handled differently by this library. This was not the
case and the same error occured. Continuing to work with
PIL Image as the code appeared more readable I then implemented
a function called validate_image_size. This excludes images of
size zero from being saved and throwing the same exception.

As a result of implementing this function 20 images are lost from
the training set and 4 images are lost from the validation set.
While we want as many images as possible this is not a major loss
and should have no noticable negative impact on the training of
our final model.