Add function to preprocess images from directories
Function written to use process_image function while it iterates
through the directories created during the image extraction phase.

Function then saves the images to a new location so that the
originals are not overwritten in the event that we want to use
different preprocessing techniques after running the model to
help achieve better results.