Json Class counter implemented
Created the functionality to iterate through the json file, using
a dict to store (Key -> Catergory, Value -> TotalCount) of each
category in the json file.

Ignoring the categories 'other vehicle', 'other person', 'trailer'
as they are not needed.

Once count is complete, it is displayed as a bar chart so we can
analysis the data set.