Add model version 2
Added model version 2 with the hopes that a more complex model
containing additional convolutional layers, along with dropout
layers will improve our accuracy.

Preprocessing stage is not excluding images below 50 pixels as
we do not want to blow these images up because quality and
information is lost during this process.

Image is now being normalised so that a common scale is being used
when training the model. The models histogram is also being
equalised to potentially enhance darker images.